+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/ChatMessageListCellFactory.java
@@ -24,10 +24,10 @@ import bisq.desktop.main.content.bisq_easy.offerbook.PeerOfferMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.MyProtocolLogMessageBox;
 import bisq.desktop.main.content.bisq_easy.open_trades.PeerProtocolLogMessageBox;
 import bisq.desktop.main.content.components.chat_messages.ChatMessageListItem;
-import bisq.desktop.main.content.components.chat_messages.list_view.messages.LeaveChatMessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.messages.MessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.messages.MyMessageBox;
-import bisq.desktop.main.content.components.chat_messages.list_view.messages.PeerMessageBox;
+import bisq.desktop.main.content.components.chat_messages.list_view.message_box.LeaveChatMessageBox;
+import bisq.desktop.main.content.components.chat_messages.list_view.message_box.MessageBox;
+import bisq.desktop.main.content.components.chat_messages.list_view.message_box.MyMessageBox;
+import bisq.desktop.main.content.components.chat_messages.list_view.message_box.PeerMessageBox;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
