+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/message_box/BubbleMessageBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chat_messages.list_view.messages;
+package bisq.desktop.main.content.components.chat_messages.list_view.message_box;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
